pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vault-mcp"
version = "0.4.0"
description = "MCP-compliant Obsidian documentation server with RAG and live sync"
authors = [{name = "Vault MCP Team", email = "team@vault-mcp.dev"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "watchdog>=3.0.0",
    "mistune>=3.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=5.1.0",
    "toml>=0.10.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "litellm>=1.35.2",
    "openai>=1.30.1",
    "mlx-lm>=0.12.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-litellm",
    "llama-index-llms-openai",
    "llama-index-vector-stores-chroma",
    "llama-index-readers-file>=0.5.0",
    "llama-index-readers-obsidian>=0.6.0",
    "llama-index-readers-joplin>=0.5.0",
    "fastapi-mcp>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
    "pyfakefs>=5.0.0",
]

[dependency-groups]
dev = [
    "pyfakefs>=5.9.2",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "types-toml>=0.10.8.20240310",
]


[project.scripts]
vault-mcp = "vault_mcp.main:run"

#-- Tooling --#

[tool.hatch.build.targets.wheel]
packages = ["components", "shared"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["components", "shared", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "SIM", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG002", "F841", "SIM117", "B018"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]
namespace_packages = true
explicit_package_bases = true

[tool.bandit]
exclude_dirs = [
    "tests", 
    "components/agentic_retriever/tests", 
    "components/api_server/tests", 
    "components/document_processing/tests", 
    "components/embedding_system/tests", 
    "components/file_watcher/tests", 
    "components/mcp_server/tests", 
    "components/vector_store/tests"
]
skips = ["B101"]

[tool.pytest.ini_options]
timeout = 30
timeout_method = "thread"
testpaths = ["tests", "components/*/tests"]
python_files = ["test_*.py"]
addopts = "--cov=vault_mcp --cov=components --cov-report=term-missing"
asyncio_default_fixture_loop_scope = "function"
